Switch from operationcount to maxexecutiontime
Clients living on a different chassis than the servers take longer to finish
the same number of operations, since their latency is higher, which can skew
the throughput measurements.

If we do not specify a high operationcount, the client will use the default
operationcount of 1000 and run for less than second, since it will use the
earlier of maxexecutiontime and operationcount.